refactor: data 패키지 통합
