+++ src/main/java/com/bookspot/batch/step/StockStepConfig.java
@@ -1,9 +1,9 @@
 package com.bookspot.batch.step;
 
-import com.bookspot.batch.stock.data.CurrentLibrary;
-import com.bookspot.batch.stock.data.LibraryStock;
-import com.bookspot.batch.stock.data.LibraryStockCsvData;
-import com.bookspot.batch.stock.data.StockFileData;
+import com.bookspot.batch.data.LibraryForFileParsing;
+import com.bookspot.batch.data.LibraryStock;
+import com.bookspot.batch.data.file.csv.LibraryStockCsvData;
+import com.bookspot.batch.data.crawler.StockFileData;
 import com.bookspot.batch.step.processor.csv.stock.LibraryStockProcessor;
 import com.bookspot.batch.step.processor.crawler.StockFilePathParser;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +28,7 @@ public class StockStepConfig {
     private final LibraryStockProcessor LibraryStockProcessor;
     private final JdbcBatchItemWriter<LibraryStock> libraryStockWriter;
 
-    private final JdbcPagingItemReader<CurrentLibrary> libraryStockDataReader;
+    private final JdbcPagingItemReader<LibraryForFileParsing> libraryStockDataReader;
     private final StockFilePathParser stockFilePathParser;
     private final ItemWriter<StockFileData> stockFileDownloaderWriter;
 
@@ -45,7 +45,7 @@ public class StockStepConfig {
     @Bean
     public Step stockFileDownloadStep() {
         return new StepBuilder(StockStepConst.DOWNLOAD_STEP_NAME, jobRepository)
-                .<CurrentLibrary, StockFileData>chunk(StockStepConst.DOWNLOAD_CHUNK_SIZE, platformTransactionManager)
+                .<LibraryForFileParsing, StockFileData>chunk(StockStepConst.DOWNLOAD_CHUNK_SIZE, platformTransactionManager)
                 .reader(libraryStockDataReader)
                 .processor(stockFilePathParser)
                 .writer(stockFileDownloaderWriter)
