+++ src/main/java/com/bookspot/batch/step/processor/csv/stock/LibraryStockProcessor.java
@@ -2,8 +2,8 @@ package com.bookspot.batch.step.processor.csv.stock;
 
 import com.bookspot.batch.step.processor.csv.stock.repository.BookRepository;
 import com.bookspot.batch.step.processor.csv.stock.repository.LibraryRepository;
-import com.bookspot.batch.stock.data.LibraryStock;
-import com.bookspot.batch.stock.data.LibraryStockCsvData;
+import com.bookspot.batch.data.LibraryStock;
+import com.bookspot.batch.data.file.csv.LibraryStockCsvData;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemProcessor;
