+++ src/main/java/com/bookspot/batch/step/reader/LibraryExcelReaderConfig.java
@@ -1,6 +1,6 @@
 package com.bookspot.batch.step.reader;
 
-import com.bookspot.batch.library.data.Library;
+import com.bookspot.batch.data.Library;
 import com.bookspot.batch.step.reader.file.excel.library.LibraryExcelRowMapper;
 import org.springframework.batch.extensions.excel.poi.PoiItemReader;
 import org.springframework.context.annotation.Bean;
