+++ src/main/java/com/bookspot/batch/step/reader/StockCsvFileReaderConfig.java
@@ -4,7 +4,7 @@ import com.bookspot.batch.step.processor.csv.book.YearParser;
 import com.bookspot.batch.step.reader.file.csv.stock.StockCsvDataMapper;
 import com.bookspot.batch.step.reader.file.csv.stock.StockCsvDelimiterTokenizer;
 import com.bookspot.batch.step.writer.file.stock.StockCsvMetadataCreator;
-import com.bookspot.batch.stock.data.LibraryStockCsvData;
+import com.bookspot.batch.data.file.csv.LibraryStockCsvData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.item.file.FlatFileItemReader;
