+++ src/main/java/com/bookspot/batch/step/writer/file/stock/StockFileDownloader.java
@@ -1,7 +1,7 @@
 package com.bookspot.batch.step.writer.file.stock;
 
 import com.bookspot.batch.global.file.NaruFileDownloader;
-import com.bookspot.batch.stock.data.StockFileData;
+import com.bookspot.batch.data.crawler.StockFileData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
